Fix non-tapable arcTap
Change some of the generation rules.

When the arc start and end was the same time.
It will bugged the arcTap and display the strange line in the game in Arcade-One.

This bug fixed when the arc start and end is the same time. It will add one more ms into the end and the start and end axis will be the same to avoid the line fly away.